fix: add start:multi script & sort files
